Proper shutdown of the JDBC drivers
Deregister all JDBC drivers loaded by the WebappClassloader on context
shutdown to avoid a memory leak.

Special handling is needed for the mysql JDBC driver as the
AbandonedConnectionCleanupThread must be shut down manually before the
driver is deregistered.

The whole shutdown process is done with reflections to avoid dependencies
to non existing JDBC drivers.