Provide a way to retrieve the patient's data as jsonp
This extends the patients resource to provide the patient's data as jsonp
with a custom callback function.
The URI is /patients/jsonp?tokenId={tid}&callback={callback}. Though
javascript isn't that restrictive, the callback must match the regex
pattern /[a-zA-Z][a-zA-Z0-9_]*/ .
The method that generates the patient list does not generate the output
itself anymore to be more generic.