Fix wrong type mapping for double when using HSQL
OpenJPA maps Java type double to SQL type NUMERIC by default, which
causes an error when inserting Double.NEGATIVE_INFINITY (this is set as
best matching weight for the first patient entered into the database).
This fix overrides the default mapping only when HSQL is actually used
instead of setting a general mapping via @ColumnDefinition (this was the
former solution) in order to avoid possible problems with other
databases.