Persist IDGeneratorMemories with IDRequest
Previously each IDGenerator had its own transaction for updating the counter (and other properties) of the memory in the database. This is not necessary for a new patient and therefore the entities are persisted in the same transaction as the IDRequest.
If an ID is requested lazily, a separate commit is still necessary.